feat: apply automocking algorithm
